New ind (index) type
The reason that var can't function as ind is that var cannot have an implicit
conversion operator long int() or whatever.  This in turn is because other var
operators become ambigious.  In the case of ind, operator intptr_t() is key -
ind behaves as a long int if at all possible, and doesn't have ambiguous
operators.