Add a namespace; fix asum
Two commits that shouldn't be blurred together, but hey.  Namespace is libvar
for the time being; maybe for ever.  ASum was tricky because the output size is
not the input size, and it's just a var, hence the array() stuff to ensure
something is an array and can be referenced.