Write sndfiles; switch the attributes to vfile args
The attributes of a sndfile should persist from read to write.  This changes
the semantic of sndfiles a bit, but with some general implication for the
semantics of files in general.  The other way to do it would be to pass some
kind of record like an XML element when writing sndfiles, but it seems onerous.