Modify some basic accessors to use recursive dereferencing
So, dereference() is quite heavy.  This changes the behaviour so that, instead
of a deref, the accessors check for reference and call a partner function in
the Heap class.  That in turn either handles it or calls the original function
recursively.  This means that no new var is ever created, so it's much faster.