Remove lube directory from the include path for tests
This was causing - I guess - something in the MacOS stdlib to find lube/string.h instead of the system one.  A repercussion is that tests should include lube/lube.h instead of just lube.h, which also ensures that they find the one in the build tree, not an installed one.